[FIX] le build n'était pas possible

+++ Code/Projet_Barman/app/src/main/java/modele/Niveau.java
@@ -54,8 +54,8 @@ public class Niveau {
      * @param nbMaxPoints
      * @param tpsshake
      */
-    public void pts(int tpsjoueur, int nbMaxPoints, int tpsshake){
-        int pts=(tpsjoueur*nbMaxPoints)/tpsshake;
+    public void pts(int tpsjoueur, int nbMaxPoints, long tpsshake){
+        int pts=(tpsjoueur*nbMaxPoints)/(int)tpsshake;
         setPtsJoueur(pts);
         victoire();
     }
